+++ pages/api/user/[user]/contributions/index.js
@@ -42,8 +42,12 @@ export default async function (req, res) {
               }
             }
           }
-          repositoryContributions(last: 10) {
+          repositoryContributions(first: 10) {
             totalCount
+            pageInfo {
+              hasNextPage
+              endCursor
+            }
             nodes {
               repository {
                 name
@@ -63,8 +67,7 @@ export default async function (req, res) {
           }
         }
       }
-    }
-  `,
+    }`,
   };
 
   const data = await fetchGQL(gql);